# openapi/openapi.yaml
openapi: 3.1.1
info:
  title: oneSCM Handbook Studio API
  version: 0.12.0
  description: |
    Revisionssichere Erstellung rollen- und spartenspezifischer Benutzerhandbücher aus Markdown-Quellen.
    Jede Operation referenziert über `x-requirements` die zugehörigen User Stories (US-020).
    Operationen mit `x-extension: true` ergänzen die verbindliche Operationsliste aus Masterprompt §10 (siehe docs/01-gap-analyse.md, Lücke 10).
    Fehler werden als `application/problem+json` (RFC 9457) geliefert.
    Authentifizierung (ENTSCHEIDUNG E-15), je nach Betriebsart (`GET /auth/config`):
    - `oidc`: `Authorization: Bearer <JWT>` eines OpenID-Connect-Providers; technische Berechtigungen aus einem Gruppen-/Rollen-Claim.
    - `demo`: Header `X-User-Id` mit einem Demo-Benutzer; ohne Header nur Lesezugriff. Nur für Entwicklung und Demo.
    Projekt (ADR-014): Header `X-Project-Id` wählt das Projekt aller projektbezogenen Endpunkte (Standard `p_default`);
    IDs anderer Projekte werden wie nicht vorhandene behandelt (404). Ausnahmen: `/projects…` und die Health-Checks.
    Betrieb (ADR-015, ADR-027): jede Antwort trägt `X-Request-Id` (bei aktivem Tracing zusätzlich `X-Trace-Id`, W3C `traceparent` wird übernommen);
    Rate-Limits je Benutzer und Minute (429 mit `Retry-After`), mit `RATE_LIMIT_STORE=db` gemeinsam über alle Instanzen;
    Prometheus-Metriken unter `/metrics` (außerhalb von `/api/v1`, Bearer `METRICS_TOKEN` oder Berechtigung admin).
servers:
  - url: /api/v1
...
components:
  securitySchemes:
    oidcBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Access-Token des konfigurierten OIDC-Providers (AUTH_MODE=oidc)
    demoUser:
      type: apiKey
      in: header
      name: X-User-Id
...
  responses:
...
    Unauthorized:
      description: Nicht angemeldet oder Token ungültig
      headers:
        WWW-Authenticate: { schema: { type: string, example: Bearer } }
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Problem:
      description: Fehler (RFC 9457)
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
  schemas:
...
    Problem:
      type: object
      required: [type, title, status]
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string }
        gate: { $ref: '#/components/schemas/GateResult' }
        blockers: { type: array, items: { type: object } }
    RoleCode: { type: string, enum: [dealer, market, mo, hq, all] }
    DivisionCode: { type: string, enum: [car, van, truck, bus, all, unconfirmed] }
...
    GateResult:
      type: object
      properties:
        passed: { type: boolean }
        checks:
          type: array
          items:
            type: object
            properties:
              code: { type: string }
              label: { type: string }
              passed: { type: boolean }
              details: { type: array, items: { type: string } }
    ExportRequest:
      type: object
      properties:
        chapterIds: { type: array, items: { type: string } }
        roles: { type: array, items: { $ref: '#/components/schemas/RoleCode' } }
        divisions: { type: array, items: { $ref: '#/components/schemas/DivisionCode' } }
        market: { type: [string, 'null'] }
        release: { type: [string, 'null'] }
        format: { type: string, enum: [md, html, pdf, json], default: md }
        outlineId: { type: string, description: 'Handbuch-Variante (ADR-034): freigegebene Kapitel dieser Gliederung, Filter aus ihrer Variante' }
        appendices: { type: boolean, description: 'Verzeichnisse (Abkürzungen, Glossar, Bildverzeichnis, FAQ) anhängen; Standard bei Varianten: true' }
security:
  - oidcBearer: []
  - demoUser: []
paths:
...
  /optimizations:
    get:
      operationId: getOptimizations
      summary: Kennzahlen je Kapitel und priorisierte Handlungsempfehlungen
      tags: [System]
      x-requirements: [US-013]
      x-extension: true
      responses:
        '401': { $ref: '#/components/responses/Unauthorized' }
        '200': { description: Übersicht }
...
  /exports:
    post:
      operationId: createExport
      summary: Freigegebene Kapitel rollen-/sparten-/markt-/releasegefiltert exportieren
      tags: [Export]
      x-requirements: [US-010, US-012, US-014]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/ExportRequest' } } } }
      responses:
        '401': { $ref: '#/components/responses/Unauthorized' }
        '201': { description: Export erstellt }
        '409': { description: 'Blockiert (offene Blocker-/Datenschutzbefunde), content: { application/problem+json: { schema: { $ref: ''#/components/schemas/Problem'' } } }' }
    get:
      operationId: listExports
      summary: Exporte auflisten
      tags: [Export]
      x-requirements: [US-014]
      x-extension: true
      responses:
        '401': { $ref: '#/components/responses/Unauthorized' }
        '200': { description: Exporte }
  /exports/{exportId}/download:
    get:
      operationId: downloadExport
      summary: Exportdatei herunterladen
      tags: [Export]
      x-requirements: [US-014]
      x-extension: true
      parameters: [{ name: exportId, in: path, required: true, schema: { type: string } }]
      responses:
        '401': { $ref: '#/components/responses/Unauthorized' }
        '200': { description: Datei }
  /traceability:
    get:
      operationId: getTraceability
      summary: Traceability-Matrix als JSON, CSV, Markdown oder Excel
      tags: [Traceability]
      x-requirements: [US-020]
      parameters: [{ name: format, in: query, schema: { type: string, enum: [json, csv, md, xlsx], default: json } }]
      responses:
        '401': { $ref: '#/components/responses/Unauthorized' }
        '200': { description: Matrix }
  /settings:
    get:
      operationId: getSettings
      summary: Einstellungen (Schwellenwerte, Uploadgrenzen, Terminologie)
      tags: [System]
      x-requirements: [US-005, US-001, US-015]
      x-extension: true
      responses:
        '401': { $ref: '#/components/responses/Unauthorized' }
        '200': { description: Einstellungen }
    put:
      operationId: putSettings
      summary: Einstellungen ändern (Berechtigung admin)
      tags: [System]
      x-requirements: [US-005, US-001, US-015]
      x-extension: true
      requestBody: { required: true, content: { application/json: { schema: { type: object } } } }
      responses:
        '401': { $ref: '#/components/responses/Unauthorized' }
        '200': { description: Einstellungen }
        '403': { $ref: '#/components/responses/Problem' }
  /reference:
    get:
      operationId: getReference
      summary: Referenzdaten (Rollen, Sparten, Icons, Status, Entscheidungsoptionen)
      tags: [System]
      x-requirements: [US-003, US-004, US-010]
      x-extension: true
      responses:
        '401': { $ref: '#/components/responses/Unauthorized' }
        '200': { description: Referenzdaten }
  /me:
    get:
      operationId: getMe
      summary: Aktueller Benutzer mit technischen Berechtigungen
      tags: [System]
      x-requirements: [US-003]
      x-extension: true
      responses:
        '401': { $ref: '#/components/responses/Unauthorized' }
        '200': { description: Benutzer }
  /dashboard:
    get:
      operationId: getDashboard
      summary: Kennzahlen für Dashboard und Optimierungen
      tags: [System]
      x-requirements: [US-013]
      x-extension: true
      responses:
        '401': { $ref: '#/components/responses/Unauthorized' }
        '200': { description: Kennzahlen }
  /audit-events:
    get:
      operationId: listAuditEvents
      summary: Auditprotokoll
      tags: [System]
      x-requirements: [US-007, US-009, US-012]
      x-extension: true
      parameters:
        - { name: entityType, in: query, schema: { type: string } }
        - { name: entityId, in: query, schema: { type: string } }
        - { name: limit, in: query, schema: { type: integer } }
      responses:
        '401': { $ref: '#/components/responses/Unauthorized' }
        '200': { description: Ereignisse }
...
  /search:
    get:
      operationId: globalSearch
      summary: 'Globale Suche über Kapitel, Kapiteltexte, Textschnipsel, Quellen, Gliederungen, Abkürzungen, Glossar und FAQ (ADR-035)'
      tags: [Suche]
      x-requirements: [NFR-22]
      x-extension: true
      parameters:
        - { name: q, in: query, required: true, schema: { type: string, minLength: 2, maxLength: 200 } }
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 25, default: 8 }, description: 'Treffer je Bereich' }
      responses:
        '401': { $ref: '#/components/responses/Unauthorized' }
        '200': { description: '{ query, total, groups: [{ type, label, hits: [{ type, id, title, excerpt, link }] }] }' }
        '400': { description: 'Suchbegriff zu kurz oder zu lang' }
...